requirements.txt
pandas
numpy
pydantic
openai
openenv-core
python-dotenv